requirements.txt
Django==4.2.16
 djangorestframework==3.15.2
 psycopg2-binary==2.9.9
 gunicorn==21.2.0
 whitenoise==6.7.0
 django-environ==0.11.2
 stripe==10.2.0
 celery==5.4.0
 redis==5.0.8
 openai==1.46.0
 pillow==10.4.0
 requests==2.32.3
 python-dateutil==2.9.0.post0
 twilio==9.3.2
 openpyxl==3.1.5
 django-cors-headers==4.4.0
